+++ nodeconductor/logging/tests/factories.py
@@ -99,3 +99,22 @@ class WebHookFactory(factory.DjangoModelFactory):
         if hook is None:
             hook = WebHookFactory()
         return 'http://testserver' + reverse('webhook-detail', kwargs={'uuid': hook.uuid})
+
+
+class PushHookFactory(factory.DjangoModelFactory):
+    class Meta:
+        model = models.PushHook
+
+    event_types = get_valid_events()[:3]
+    token = 'VALID_TOKEN'
+    type = models.PushHook.Type.ANDROID
+
+    @classmethod
+    def get_list_url(self):
+        return 'http://testserver' + reverse('pushhook-list')
+
+    @classmethod
+    def get_url(self, hook=None):
+        if hook is None:
+            hook = PushHookFactory()
+        return 'http://testserver' + reverse('pushhook-detail', kwargs={'uuid': hook.uuid})