+++ nodeconductor/structure/apps.py
@@ -149,6 +149,12 @@ class StructureConfig(AppConfig):
             dispatch_uid='nodeconductor.structure.handlers.log_project_group_role_revoked',
         )
 
+        signals.pre_delete.connect(
+            handlers.revoke_roles_on_project_deletion,
+            sender=Project,
+            dispatch_uid='nodeconductor.structure.handlers.revoke_roles_on_project_deletion',
+        )
+
         for index, model in enumerate(ResourceMixin.get_all_models()):
             signals.pre_delete.connect(
                 handlers.log_resource_deleted,
@@ -199,50 +205,6 @@ class StructureConfig(AppConfig):
                     dispatch_uid='nodeconductor.structure.handlers.detect_vm_coordinates_{}_{}'.format(
                         model.__name__, index),
                 )
-
-        for model in ServiceProjectLink.get_all_models():
-            name = 'propagate_ssh_keys_for_%s' % model.__name__
-            signals.post_save.connect(
-                handlers.propagate_user_to_his_projects_services,
-                sender=model,
-                dispatch_uid='nodeconductor.structure.handlers.%s' % name,
-            )
-
-        signals.pre_delete.connect(
-            handlers.remove_stale_user_from_his_projects_services,
-            sender=User,
-            dispatch_uid='nodeconductor.structure.handlers.remove_stale_user_from_his_projects_services',
-        )
-
-        signals.post_save.connect(
-            handlers.propagate_new_users_key_to_his_projects_services,
-            sender=SshPublicKey,
-            dispatch_uid='nodeconductor.structure.handlers.propagate_new_users_key_to_his_projects_services',
-        )
-
-        signals.post_delete.connect(
-            handlers.remove_stale_users_key_from_his_projects_services,
-            sender=SshPublicKey,
-            dispatch_uid='nodeconductor.structure.handlers.remove_stale_users_key_from_his_projects_services',
-        )
-
-        signals.pre_delete.connect(
-            handlers.revoke_roles_on_project_deletion,
-            sender=Project,
-            dispatch_uid='nodeconductor.structure.handlers.revoke_roles_on_project_deletion',
-        )
-
-        structure_signals.structure_role_granted.connect(
-            handlers.propagate_user_to_services_of_newly_granted_project,
-            sender=Project,
-            dispatch_uid='nodeconductor.structure.handlers.propagate_user_to_services_of_newly_granted_project',
-        )
-
-        structure_signals.structure_role_revoked.connect(
-            handlers.remove_stale_user_from_services_of_revoked_project,
-            sender=Project,
-            dispatch_uid='nodeconductor.structure.handlers.remove_stale_user_from_services_of_revoked_project',
-        )
 
         structure_signals.customer_account_credited.connect(
             handlers.log_customer_account_credited,